refactor generator into utils header, add more tests
